Add files via upload
This C++ program demonstrates a simple swapping technique for two integer variables, a and b, using a temporary variable temp. Initially, it displays the original values of a (5) and b (10). To swap the values, the program uses temp as a holding variable. First, temp is assigned the value of a (5), then a is assigned the value of b (10), and finally, b is assigned the value stored in temp (5). This effectively exchanges the values of a and b. The program then displays the swapped values, showing that a now holds 10 and b holds 5.