Add files via upload
This C++ program calculates and displays the square root of a given number using the sqrt function from the cmath library. The program begins by including the necessary libraries: cmath for mathematical functions and iostream for input/output operations. The using namespace std; directive allows standard library functions to be used without prefixing std::. In the main function, a double-precision floating-point variable num is initialized with the value 100.0. The sqrt function is then used to calculate the square root of num, and the result is output to the console using cout. The program terminates by returning an exit status of 0. When run, the program will output: "Square root of 100 is: 10", which is the result of calculating the square root of 100 using the sqrt function. This program demonstrates the use of mathematical functions from the cmath library and basic input/output operations in C++