Add files via upload
This C++ program calculates and displays the result of raising a number to a specified power using the pow function from the cmath library. The program includes the necessary libraries, cmath for mathematical functions and iostream for input/output operations. The using namespace std; directive enables the use of standard library functions without prefixing std::. In the main function, the pow function is utilized to calculate 4 raised to the power of 2. The pow function takes two arguments: the base (4) and the exponent (2). The result is then output to the console using cout. The program terminates by returning an exit status of 0. When executed, the program displays: "4 raised to the power 2 is: 16", demonstrating the successful calculation of 4 squared. This program showcases the application of mathematical functions from the cmath library and basic input/output operations in C++.