Add files via upload
This C++ program demonstrates pass-by-value for strings, where a copy of the original string is passed to the printString function. In the main function, a string variable myString is initialized with "Good morning" and passed to printString. Within printString, a local copy of myString is created and printed to the console using cout. The program outputs "Good morning" to the console, illustrating a simple example of function usage and string passing