Add files via upload
This C++ program demonstrates function using pointers, where a pointer to an integer variable x is passed to the fun function, allowing modification of the original variable. In main, x is initialized to 35, and its address is passed to fun using the unary & operator. Within fun, the dereference operator  is used to modify the value at the passed address, setting x to 15. After fun returns, the modified value of x is printed to the console, outputting "x = 15"