Add files via upload
This C++ program defines a Rectangle class that demonstrates the use of both a default constructor and a parameterized constructor. The class has private attributes length and breadth. The default constructor initializes these attributes to zero, while the parameterized constructor allows the user to specify their values upon object creation. In the main function, an object rect1 is created using the default constructor, resulting in a rectangle with dimensions of 0 for both length and breadth. The program then displays these dimensions and calculates the area, which will also be 0. Next, an object rect2 is created using the parameterized constructor with dimensions of 10 and 5. The dimensions and area of rect2 are then displayed. This program effectively illustrates how constructors can be used to initialize objects in different ways, providing flexibility in object creation within the same class.