Add files via upload
This C++ program defines a Time class that illustrates the use of both a default constructor and a parameterized constructor for initializing time values. The class contains private attributes for hours, minutes, and seconds. The default constructor initializes these attributes to zero, while the parameterized constructor allows users to set specific values for hours, minutes, and seconds upon creating an object. In the main function, an object t1 is instantiated using the default constructor, which results in a time of 0:0:0, and this is displayed. Another object, t2, is created using the parameterized constructor with the values 10 hours, 45 minutes, and 30 seconds. The program then displays this custom time. Overall, this program effectively demonstrates how constructors can be utilized to create and initialize objects in different ways, providing flexibility in managing time representations.