Add files via upload
This C++ program defines a simple class named Test that demonstrates the execution of a user-defined constructor and destructor. The constructor outputs the message "Constructor executed" when an instance of the class is created, indicating that the object is being initialized. Conversely, the destructor outputs "Destructor executed" when the object goes out of scope and is destroyed, signaling the cleanup process. In the main function, an instance of Test named t is created. As the program runs, it will first display the constructor message when t is instantiated, and then, when the program exits the scope of main, the destructor message will be displayed. This program effectively illustrates the lifecycle of an object in C++, highlighting the roles of constructors and destructors in resource management.