Add files via upload
This C++ program demonstrates the concept of polymorphism through the use of function overriding in a base and derived class. The Base class contains a function overriding  named print(), which outputs "Base Function." The Derived class, inheriting from Base, overrides this method with its own implementation that outputs "Derived Function." In the main function, an instance of Derived called derived1 is created. When the print() method is called on this instance, it executes the overridden version in the Derived class, resulting in the output "Derived Function." This behavior illustrates how polymorphism allows derived classes to provide specific implementations for methods defined in base classes, enabling more flexible and dynamic code execution.