Add files via upload
This C++ program demonstrates how to extract a substring from a string using the std::string class. It initializes a string variable text with the value "Hello World" and then uses the substr method to extract a portion of the string. Specifically, text.substr(6, 5) retrieves a substring starting at index 6 (which corresponds to the letter 'W' in "World") and takes the next 5 characters. The extracted substring, "World", is stored in the variable sub. Finally, the program outputs the extracted substring with the message "Substring: World". This example effectively illustrates how to use the substr method to manipulate and retrieve specific portions of a string in C++.