Add files via upload
This C++ program demonstrates the use of the std::string constructor that creates a string initialized with a specified number of repeated characters. The line string str(5, 'g'); creates a string variable str containing five consecutive 'g' characters, resulting in the string "ggggg". The program then outputs this string using cout, which will display "ggggg" on the console. This example effectively illustrates how to initialize a string with repeated characters in C++.