Add files via upload
This C++ program defines a simple class named MyClass, which includes two public attributes: an integer myNum and a string myString. The main function creates an instance of MyClass called myObj. The program then sets the values of myNum to 20 and myString to "some text" using dot notation to access the attributes of the object. Finally, it outputs these values to the console, printing 20 followed by some text. This example illustrates the basic concepts of class definition, object creation, and attribute access in C++.